requirements.txt
# Core requirements
python-dotenv==1.0.0
pydantic==2.5.2
litellm==1.10.0
pytest==7.4.3
jupyter==1.0.0
matplotlib==3.8.2
pandas==2.1.3

# Framework-specific
langchain==0.1.0
langgraph==0.1.0
dspy-ai==2.3.0
google-adk==0.0.2
inspect-ai==1.0.0
smolagents==0.1.0
pydantic-ai==0.1.0
agno==0.1.0
